+++ src/main/java/zaftnotameni/creatania/commands/CreataniaTesterCommand.java
@@ -1,38 +1,43 @@
 package zaftnotameni.creatania.commands;
-import com.mojang.authlib.GameProfile;
 import com.mojang.brigadier.CommandDispatcher;
+import com.mojang.brigadier.arguments.IntegerArgumentType;
 import com.mojang.brigadier.context.CommandContext;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
-import net.minecraft.core.Direction;
+import net.minecraft.commands.arguments.blocks.BlockStateArgument;
+import net.minecraft.network.chat.TextComponent;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.world.InteractionHand;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
-import net.minecraft.world.phys.BlockHitResult;
-import net.minecraft.world.phys.Vec3;
-import net.minecraftforge.common.util.FakePlayer;
 
 import java.util.UUID;
 public class CreataniaTesterCommand {
   public CreataniaTesterCommand(CommandDispatcher<CommandSourceStack> dispatcher) {
-    dispatcher.register(Commands.literal("creatania").then(Commands.literal("test").executes(command -> {
-      return creataniaTest(command);
-    })));
+    dispatcher.register(
+      Commands.literal("creatania").then(
+        Commands.literal("checkerboardfloor").then(
+          Commands.argument("range", IntegerArgumentType.integer(1, 300)).then(
+            Commands.argument("block1", BlockStateArgument.block()).then(
+              Commands.argument("block2", BlockStateArgument.block())
+                .executes(CreataniaTesterCommand::creataniaTest))))));
   }
 
   public static int creataniaTest(CommandContext<CommandSourceStack> command) {
     var source = command.getSource();
+    var range = IntegerArgumentType.getInteger(command, "range");
+    var b1 = BlockStateArgument.getBlock(command, "block1").getState();
+    var b2 = BlockStateArgument.getBlock(command, "block2").getState();
     var player = source.getEntity();
     var level = source.getLevel();
     if (level == null || level.isClientSide || player == null || !(player instanceof ServerPlayer p)) return 0;
-    var fakePlayer = new FakePlayer(level, new GameProfile(UUID.randomUUID(), "fake"));
-    fakePlayer.getInventory().add(new ItemStack(Items.COAL_BLOCK, 64));
-    var pos = p.getOnPos().west().above();
-    var state = level.getBlockState(pos);
-    var block = state.getBlock();
-    var hit = new BlockHitResult(new Vec3(pos.getX(), pos.getY(), pos.getZ()), Direction.NORTH, pos, false);
-    block.use(state, level, pos, fakePlayer, InteractionHand.MAIN_HAND, hit);
+    var pos = player.getOnPos();
+    player.sendMessage(new TextComponent("checkerboard started"), UUID.randomUUID());
+    for (var x = -range; x <= range; x++) for (var z = -range; z <= range; z++) {
+      var targetPos = pos.offset(x, 0, z);
+      var b = ((x + z) % 2) == 0 ? b1 : b2;
+      player.sendMessage(new TextComponent(
+        "Setting block at " + targetPos.toShortString() + " to " + b.getBlock().getDescriptionId()), UUID.randomUUID());
+      level.setBlockAndUpdate(targetPos, b);
+    }
+    player.sendMessage(new TextComponent("checkerboard finished"), UUID.randomUUID());
     return 0;
   }
 }