+++ src/main/java/zaftnotameni/creatania/machines/manamotor/ManaMotorBlockEntity.java
@@ -2,11 +2,13 @@ package zaftnotameni.creatania.machines.manamotor;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.simibubi.create.content.contraptions.base.GeneratingKineticTileEntity;
+import com.simibubi.create.foundation.config.AllConfigs;
 import com.simibubi.create.foundation.tileEntity.TileEntityBehaviour;
 import com.simibubi.create.foundation.tileEntity.behaviour.scrollvalue.ScrollValueBehaviour;
 import net.minecraft.client.Minecraft;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.ItemStack;
@@ -33,6 +35,8 @@ import zaftnotameni.sharedbehaviors.KineticManaMachine;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.List;
+
+import static zaftnotameni.creatania.util.Text.*;
 
 /**
  * Generates SU (from Create) when provided Mana (from Botania).
@@ -94,6 +98,30 @@ public class ManaMotorBlockEntity extends GeneratingKineticTileEntity implements
       .withStressUnitsPerRpm(ManaMotorConfig.getStressUnitsPerRPM())
       .withBaseRpm(CommonConfig.MANA_MOTOR_BASE_RPM.get());
      return this.manaMachine;
+  }
+  @Override
+  public boolean addToGoggleTooltip(List<Component> tooltip, boolean isPlayerSneaking) {
+    super.addToGoggleTooltip(tooltip, isPlayerSneaking);
+
+    purple("").forGoggles(tooltip);
+
+    muted("Stress Units produced per RPM:").space()
+      .add(gray(String.valueOf(this.getManaMachine().stressUnitsPerRpm))).forGoggles(tooltip);
+
+    purple("").forGoggles(tooltip);
+
+    muted("Maximum SU Produced:").space()
+      .add(gray(String.valueOf(this.getManaMachine().getMaximumSUPossible()))).space()
+      .add(muted("at")).space()
+      .add(gray(String.valueOf(AllConfigs.SERVER.kinetics.maxMotorSpeed.get()))).space()
+      .add(gray("RPM")).forGoggles(tooltip);
+
+    purple("").forGoggles(tooltip);
+
+    muted("Mana consumed per RPM:").space()
+      .add(red("HIGH")).forGoggles(tooltip);
+
+    return true;
   }
   @Override
   public void receiveMana(int pMana) { this.getManaMachine().receiveMana(pMana); }